+++ media/smart-card/src/view/Card/index.tsx
@@ -1,9 +1,13 @@
 import React from 'react';
-import { withAnalyticsEvents } from '@atlaskit/analytics-next';
+import {
+  withAnalyticsContext,
+  withAnalyticsEvents,
+} from '@atlaskit/analytics-next';
 import { CardAppearance, CardPlatform, CardProps } from './types';
 import { CardWithDataRenderer } from '../CardWithData/loader';
 import { CardWithURLRenderer } from '../CardWithUrl/loader';
 import { isCardWithData } from '../../utils';
+import { context } from '../../utils/analytics';
 
 class PlainCard extends React.PureComponent<CardProps> {
   render() {
@@ -15,6 +19,8 @@ class PlainCard extends React.PureComponent<CardProps> {
   }
 }
 
-export const Card = withAnalyticsEvents()(PlainCard);
+export const Card = withAnalyticsContext(context)(
+  withAnalyticsEvents()(PlainCard),
+);
 
 export type { CardAppearance, CardProps, CardPlatform };